[FIX] auth_signup_delegate: State is better to know if user has already logged in
If a user ask to reset its password (signup_type=reset), then an error is raised when opening the delegated user creation form.
Same if a new portal access is given (signup_type=signup) using the wizard.

In both cases, the user is not required to reset its password.
The `signup_type` and `signup_token` should be cleared after a successfull login